Remove remaining monkey-patching in codebase
Also remove external access to private methods of `StatusEffectsPF2e`
