+++ .betterer.results
@@ -5,37 +5,37 @@
 //
 exports[`too-much-lint`] = {
   value: `{
-    "src/module/actor/character/sheet.ts:2136679983": [
-      [117, 72, 3, "Unexpected any. Specify a different type.", "193409811"],
-      [117, 77, 3, "Unexpected any. Specify a different type.", "193409811"],
-      [125, 72, 3, "Unexpected any. Specify a different type.", "193409811"],
-      [125, 77, 3, "Unexpected any. Specify a different type.", "193409811"],
-      [212, 64, 3, "Unexpected any. Specify a different type.", "193409811"],
-      [292, 30, 3, "Unexpected any. Specify a different type.", "193409811"]
+    "src/module/actor/character/sheet.ts:1722372315": [
+      [118, 72, 3, "Unexpected any. Specify a different type.", "193409811"],
+      [118, 77, 3, "Unexpected any. Specify a different type.", "193409811"],
+      [126, 72, 3, "Unexpected any. Specify a different type.", "193409811"],
+      [126, 77, 3, "Unexpected any. Specify a different type.", "193409811"],
+      [213, 64, 3, "Unexpected any. Specify a different type.", "193409811"],
+      [293, 30, 3, "Unexpected any. Specify a different type.", "193409811"]
     ],
     "src/module/actor/modifiers.ts:2050442843": [
       [441, 19, 3, "Unexpected any. Specify a different type.", "193409811"]
     ],
-    "src/module/actor/sheet/base.ts:3716510317": [
-      [1001, 20, 3, "Unexpected any. Specify a different type.", "193409811"]
+    "src/module/actor/sheet/base.ts:4258892170": [
+      [1016, 20, 3, "Unexpected any. Specify a different type.", "193409811"]
     ],
-    "src/module/actor/vehicle/sheet.ts:3510236265": [
+    "src/module/actor/vehicle/sheet.ts:1418399256": [
       [21, 25, 3, "Unexpected any. Specify a different type.", "193409811"],
-      [34, 38, 3, "Unexpected any. Specify a different type.", "193409811"]
+      [34, 44, 3, "Unexpected any. Specify a different type.", "193409811"]
     ],
-    "src/module/item/base.ts:4265974295": [
-      [325, 24, 3, "Unexpected any. Specify a different type.", "193409811"]
+    "src/module/item/base.ts:2183679181": [
+      [332, 24, 3, "Unexpected any. Specify a different type.", "193409811"]
     ],
-    "src/module/item/consumable/data.ts:2037414684": [
+    "src/module/item/consumable/data.ts:3570976201": [
       [29, 13, 3, "Unexpected any. Specify a different type.", "193409811"]
     ],
-    "src/module/item/sheet/base.ts:1083229076": [
+    "src/module/item/sheet/base.ts:492959271": [
       [58, 25, 3, "Unexpected any. Specify a different type.", "193409811"]
     ],
     "src/module/system/tag-selector/basic.ts:1185943401": [
       [127, 44, 3, "Unexpected any. Specify a different type.", "193409811"]
     ],
-    "src/module/system/tag-selector/resistances.ts:356951034": [
+    "src/module/system/tag-selector/resistances.ts:2878489551": [
       [34, 20, 3, "Unexpected any. Specify a different type.", "193409811"],
       [83, 71, 3, "Unexpected any. Specify a different type.", "193409811"]
     ],
@@ -46,21 +46,17 @@ exports[`too-much-lint`] = {
       [20, 20, 3, "Unexpected any. Specify a different type.", "193409811"],
       [69, 71, 3, "Unexpected any. Specify a different type.", "193409811"]
     ],
-    "src/module/system/tag-selector/weaknesses.ts:1139668259": [
+    "src/module/system/tag-selector/weaknesses.ts:2483824982": [
       [24, 20, 3, "Unexpected any. Specify a different type.", "193409811"],
       [72, 71, 3, "Unexpected any. Specify a different type.", "193409811"]
-    ],
-    "src/scripts/actor/status-effects.ts:2696049250": [
-      [414, 24, 3, "Unexpected any. Specify a different type.", "193409811"],
-      [439, 32, 3, "Unexpected any. Specify a different type.", "193409811"]
     ],
     "types/foundry/client/application/form-application/document-sheet/actor-sheet.d.ts:4046500136": [
       [8, 15, 3, "Unexpected any. Specify a different type.", "193409811"],
       [9, 14, 3, "Unexpected any. Specify a different type.", "193409811"],
       [10, 15, 3, "Unexpected any. Specify a different type.", "193409811"]
     ],
-    "types/foundry/client/collections/compendium-collection.d.ts:4081699882": [
-      [191, 23, 3, "Unexpected any. Specify a different type.", "193409811"]
+    "types/foundry/client/collections/compendium-collection.d.ts:31827583": [
+      [208, 23, 3, "Unexpected any. Specify a different type.", "193409811"]
     ],
     "types/foundry/client/documents/item.d.ts:2931579527": [
       [62, 45, 3, "Unexpected any. Specify a different type.", "193409811"]
@@ -78,7 +74,7 @@ exports[`too-much-lint`] = {
       [167, 21, 3, "Unexpected any. Specify a different type.", "193409811"],
       [305, 44, 3, "Unexpected any. Specify a different type.", "193409811"]
     ],
-    "types/foundry/client/game.d.ts:494015879": [
+    "types/foundry/client/game.d.ts:217439399": [
       [170, 32, 3, "Unexpected any. Specify a different type.", "193409811"]
     ],
     "types/foundry/client/pixi/placeables-layer/base.d.ts:3844060973": [
@@ -106,8 +102,8 @@ exports[`too-much-lint`] = {
     "types/foundry/common/abstract/data.d.ts:1438318714": [
       [22, 34, 3, "Unexpected any. Specify a different type.", "193409811"]
     ],
-    "types/foundry/common/abstract/document.d.ts:2707527115": [
-      [560, 29, 3, "Unexpected any. Specify a different type.", "193409811"]
+    "types/foundry/common/abstract/document.d.ts:3086286863": [
+      [572, 29, 3, "Unexpected any. Specify a different type.", "193409811"]
     ],
     "types/foundry/common/utils/collection.d.ts:738862421": [
       [111, 31, 3, "Unexpected any. Specify a different type.", "193409811"],
