+++ src/module/system/settings/index.ts
@@ -1,7 +1,7 @@
 import { VariantRulesSettings } from "./variant-rules";
 import { WorldClockSettings } from "./world-clock";
 import { HomebrewElements } from "./homebrew";
-import { StatusEffects } from "@scripts/actor/status-effects";
+import { StatusEffects } from "@module/canvas/status-effects";
 import { objectHasKey } from "@util";
 import { MigrationRunner } from "@module/migration/runner";
 import { AutomationSettings } from "./automation";
