+++ static/lang/en.json
@@ -2589,7 +2589,7 @@
             "statusEffectType": {
                 "blackWhite": "Black & white",
                 "default": "Full color",
-                "hint": "Determines which icon set is used for condition token effects. Thanks to errational#2007 for inspiration and Ediwir#7664 for the B&W tokens by Lorc@OpenGameArt.",
+                "hint": "Determines which icon set is used for condition token effects.",
                 "legacy": "Legacy",
                 "name": "Condition Icon Set"
             },
